Use new test syntax for DetailsAndEdit component
